добавлен класс Gift + соответствующая логика
