+++ src/main/java/ru/novikova/birthdaysGifts/entities/User.java
@@ -1,9 +1,9 @@
 package ru.novikova.birthdaysGifts.entities;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.novikova.birthdaysGifts.entities.community.CommunityMember;
+import ru.novikova.birthdaysGifts.entities.gift.Gift;
 
 import javax.persistence.*;
 import java.util.Date;
@@ -27,4 +27,11 @@ public class User {
 
     @OneToMany(mappedBy = "user")
     private List<CommunityMember> communityMemberList;
+
+    @ManyToMany
+    @JoinTable(name = "user_gift",
+            joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "gift_id")
+    )
+    private List<Gift> gifts;
 }
