--- src/main/java/ru/novikova/birthdaysGifts/entities/community/FamilyMemberType.java
@@ -1,8 +1,0 @@
-package ru.novikova.birthdaysGifts.entities.community;
-
-public enum FamilyMemberType {
-    FATHER,
-    MOTHER,
-    SISTER,
-    BROTHER
-}
