--- src/main/java/ru/novikova/birthdaysGifts/entities/community/FriendType.java
@@ -1,6 +1,0 @@
-package ru.novikova.birthdaysGifts.entities.community;
-
-public enum FriendType {
-    FRIEND,
-    BEST_FRIEND
-}
