+++ src/main/java/ru/novikova/birthdaysGifts/entities/community/family/FamilyMemberType.java
@@ -1,0 +1,8 @@
+package ru.novikova.birthdaysGifts.entities.community.family;
+
+public enum FamilyMemberType {
+    FATHER,
+    MOTHER,
+    SISTER,
+    BROTHER
+}
