+++ src/main/java/ru/novikova/birthdaysGifts/entities/community/friends/FriendType.java
@@ -1,0 +1,6 @@
+package ru.novikova.birthdaysGifts.entities.community.friends;
+
+public enum FriendType {
+    FRIEND,
+    BEST_FRIEND
+}
